PDFView State Management and Create PDF Finished: Final State management component: Create PDF done.
Creates a PDF given a name and file. the handler function fills the rest of the pdf info for the PDFView like:

id, bookmarks, progNotification, autoSavePage.
Of course it has to input that to the controller but it works with static data right now (dummy create button). We'll handle persistence and addition of those new views to the Controller and model later.